requirements.txt
flask
psycopg2-binary
sqlalchemy
python-dotenv
pytest
